developed nav bar but query not dealt with on the feed side yet
