+++ src/navigation/TopicNav.jsx
@@ -1,22 +1,62 @@
-import { MenuItem } from '@mui/material'
-import { Link, useNavigate } from 'react-router-dom'
+import { Button, Menu, MenuItem } from '@mui/material'
+import { useEffect, useState } from 'react'
+import { createSearchParams, useNavigate, useSearchParams } from 'react-router-dom'
+import { getAllTopics } from '../../api/api'
 
-const TopicNav = ({ topic}) => {
+
+const TopicNav = () => {
+  const [anchorEl, setAnchorEl] = useState(null)
+  const open = Boolean(anchorEl)
   const navigate = useNavigate()
-  const navigateToTopic = () => {
-    navigate(`/myfeed?topic=${topic.slug}`)
+  const [allTopics, setAllTopics] = useState([])
+ 
+  useEffect(() => {
+    getAllTopics().then((topics) => {
+      setAllTopics(topics)
+    })
+  }, [])
+
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget)
+  }
+  const handleClose = () => {
+    setAnchorEl(null)
   }
 
+  const navigateToTopics = (topic) => {
+    navigate({
+      pathname: '/myfeed',
+      search: createSearchParams({
+        topic,
+      }).toString(),
+    })
+  }
 
   return (
-    <MenuItem
-      onClick={() => {
-        navigateToTopic(topic)
-      }}
-    >
-      {topic.slug.charAt(0).toUpperCase() + topic.slug.slice(1)}
-    </MenuItem>
+    <>
+      <Button id="menu-button" onClick={handleClick}>
+        Topics
+      </Button>
+      <Menu
+        id="topic-menu"
+        anchorEl={anchorEl}
+        open={open}
+        onClose={handleClose}
+      >
+        {allTopics.map((topic) => {
+          return (
+            <MenuItem
+              onClick={() => {
+                navigateToTopics(topic.slug), handleClose()
+              }}
+              key={topic.slug}
+            >
+              {topic.slug.charAt(0).toUpperCase() + topic.slug.slice(1)}
+            </MenuItem>
+          )
+        })}
+      </Menu>
+    </>
   )
 }
-
 export default TopicNav
